Revert "Move to separate files."
This reverts commit 776fe2a10d9d17f2978b50ad9e1bf52e6f03a9c4.
