--- client/modules/IDE/components/Toolbar/index.js
@@ -1,1 +1,0 @@
-export { default } from './Toolbar';
